Move SupplyManager::serialize() into Supply.cpp from SerializeEmpire.cpp.
When SupplyManager becomes a pimpl implementation its details will only
be available from within Supply.cpp.
